build: fix eslint setup
This is part of the process in making this repo easy to contribute to in order to attract contributors.

The linting results are completely different, depending whether you run `yarn react:dev`or `yarn lint`. This is because craco uses a very simple internal eslint config.

Craco does not support the eslint flat config, so there is currently no way to point it at the config.

The best solution appears to be to disable craco's linting and create two sets of rules, one for the main process and one for the renderer process.

The rules I have chosen are a fairly common set for Electron Typescript React projects.

Of course, you may disagree with the rules I have chosen; just let me know... your project, your rules.

As part of this process I have also added a `.prettierrc` because eslint needs this in order for its Prettier plugins to work.