build: fix eslint setup
This is part of the process in making this repo easy to contribute to in order to attract contributors.

The linting results are completely different, depending whether you run `yarn react:dev`or `yarn lint`. This is because craco uses a very simple internal eslint config.

Craco does not support the eslint flat config, so there is currently no way to point it at the config.

The best solution appears to be to disable craco's linting and create two sets of rules, one for the main process and one for the renderer process.

The rules I have chosen are a fairly common set for Electron Typescript React projects.

Of course, you may disagree with the rules I have chosen; just let me know... your project, your rules.

As part of this process I have also added a `.prettierrc` because eslint needs this in order for its Prettier plugins to work.

The change to `craco.config.ts` is to disable craco's TypeScript error overlay.

This is one pull request in a stack... they are all based on top of each other, so the order is important.

The advantage of this approach is that you are guaranteed no conflicts and reviews should be shorter.

Review each patch and merge before you move on to the next, this way, nothing will break and the code will be easier to review.

Start from the bottom (above main) and work your way up the stack until you reach the top.

* **#25** 👈
* **main**