fix: use installed adb instead of platform-tools adb
## Problem

When there is an installed version of `adb`, no app images are displayed and no apps can be downloaded.

## Analysis

When there is an installed version of `adb`, we don't download the platform-tools, but then try to use `adb` in the `platform-tools` folder, which throws an error, breaking the app. We should use the installed version of `adb` instead.

The error is due to an async issue in the `SystemProcess` class constructor (comments to explain the problem):

```js
// /Users/username/Library/Application Support/qrookie-node/data/uncrompressed/platform-tools/adb
SystemProcess.adbPath = path.join(platformToolsDir, "adb" + binExt);

this.getCommanPath("adb") // /usr/bin/adb
  .then(path => {
    // This is the problem. Because the `getCommandPath` call is asynchronous, the constructor will continue to execute and other methods will use the initial value of `SystemProcess.adbPath` instead of `/usr/bin/adb`.
    if (path) {
      SystemProcess.adbPath = path;
      log.info("ADB found at:", path);
      return;
    }
    log.warn(`ADB not found, downloading platform-tools. getCommanPath returned: '${path}'`);
    void setupTools();
  })
```

We can't use async/await in the constructor itself, so we have to move the initialization code to an `initialize()` method, which is the solution I have chosen.