chore: rename `commands` folder to `commands`
I have created this separate from the spelling correction PR to make it easier to review and merge.

This is one pull request in a stack... they are all based on top of each other, so the order is important.

The advantage of this approach is that you are guaranteed no conflicts and reviews should be shorter.

Review each patch and merge before you move on to the next, this way, nothing will break and the code will be easier to review.

Start from the bottom (above main) and work your way up the stack until you reach the top.

* **#30** 👈
* **#29**
* **#28**
* **#22**
* **#27**
* **#26**
* **#25**
* **main**