fix: no games found container height
When the Library page contains no games, the "No Games Found" message is cut in half due to an incorrect container height.

{ screenshot }

Instead of using `height:100%`, we now calculated the height using `calc(100vh - ${navbarHeight}px - ${headerHeight}px)`.

This is one pull request in a stack... they are all based on top of each other, so the order is important.

The advantage of this approach is that you are guaranteed no conflicts and reviews should be shorter.

Review each patch and merge before you move on to the next, this way, nothing will break and the code will be easier to review.

Start from the bottom (above main) and work your way up the stack until you reach the top.

* **#31** 👈
* **#30**
* **#29**
* **#28**
* **#22**
* **#27**
* **#26**
* **#25**
* **main**